Move the SerializedSignature to cryto module
There was no good reason for putting it where it is, at least from
what I found on the original PR

> I don't remember exactly but I think I thought it wouldn't work in
> the other taproot module. I have no idea now if it's actually an
> issue

ref: https://github.com/rust-bitcoin/rust-bitcoin/pull/2156#issuecomment-1790295634

Anyways, in order to split the `crypto` module out into a separate
crate it makes sense to have the `SerializedSignature` as well.
